- last_active variable in CyAdminModel now is a Double - Now is showed the last activity label in customerly chat view controller

+++ CustomerlySDK/Library/Models/CyAdminModel.swift
@@ -12,7 +12,7 @@ class CyAdminModel: Mappable {
     var account_id : Int?
     var email : String?
     var name : String?
-    var last_active : Int?
+    var last_active : Double?
     
     
     required init?(map: Map) {
